backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
google-generativeai==0.8.3
pypdf==4.3.1
python-docx==1.1.2
python-multipart==0.0.12
